requirements.txt
astroid
attrs
certifi==2016.9.26
click==8.0.3
cycler==0.11.0
dataclasses
flask==2.0.3
importlib-metadata
iniconfig
isort
itsdangerous
Jinja2
kafka-python==2.0.2
kiwisolver==1.3.1
lazy-object-proxy
MarkupSafe
matplotlib==3.3.4
mccabe==0.6.1
more-itertools
numpy==1.19.5
packaging
Pillow==8.4.0
pluggy
py
pylint==2.7.2
pyparsing
pytest==6.2.5
python-dateutil==2.8.2
requests==2.12.5
six==1.16.0
toml
typed-ast
typing-extensions
Werkzeug
wrapt
zipp==0.6.0